requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0
cachetools==5.5.0
